AtmosphericDeposition

Documentation for AtmosphericDeposition.

AtmosphericDeposition.z₀_table

AtmosphericDeposition.DrydepositionG

AtmosphericDeposition.Wetdeposition

AtmosphericDeposition.DryDepGas

AtmosphericDeposition.DryDepParticle

AtmosphericDeposition.RbGas

AtmosphericDeposition.RbParticle

AtmosphericDeposition.WetDeposition

AtmosphericDeposition.cc

AtmosphericDeposition.dH2O

AtmosphericDeposition.dParticle

AtmosphericDeposition.mfp

AtmosphericDeposition.mu

AtmosphericDeposition.ra

AtmosphericDeposition.sc

AtmosphericDeposition.stSmooth

AtmosphericDeposition.stVeg

AtmosphericDeposition.vs

Package wesely1989 implements an algorithm for surface resistance to dry deposition.

Citation for the original article, followed by citation for an article with some corrections which have been incorporated here:

M. L. Wesely, Parameterization of surface resistances to gaseous dry deposition in regional-scale numerical models, Atmos. Environ. 23, 1293–1304 (1989), http:dx.doi.org/10.1016/0004-6981(89)90153-4.

J. Walmsley, and M. L. Wesely, Modification of coded parametrizations of surface resistances to gaseous dry deposition, Atmos. Environ. 30(7), 1181–1188 (1996), http://dx.doi.org/10.1016/1352-2310(95)00403-3.

The abstract of the original article:

Methods for estimating the dry deposition velocities of atmospheric gases in the U.S. and surrounding areas have been improved and incorporated into a revised computer code module for use in numerical models of atmospheric transport and deposition of pollutants over regional scales. The key improvement is the computation of bulk surface resistances along three distinct pathways of mass transfer to sites of deposition at the upper portions of vegetative canopies or structures, the lower portions, and the ground (or water surface). This approach replaces the previous technique of providing simple look-up tables of bulk surface resistances. With the surface resistances divided explicitly into distinct pathways, the bulk surface resistances for a large number of gases in addition IO those usually addressed in acid deposition models (SO2, O3, NOx, and HNO3) can be computed, if estimates of the effective Henry’s Law constants and appropriate measures of the chemical reactiiity of the various substances are known. This has been accomnlished successfullv for H2O2, HCHO, CH3CHO (to represent other aldehydes CH3O2H (to represent organic peroxides), CH3C(O)O2H, HCOOH (to represent organic acids), NH3, CH3C(O)O2NO2, and HNO2. Other factors considered include surface temperature, stomatal response to environmental parameters, the wetting of surfaces by dew and rain, and the covering of surfaces by snow. Surface emission of gases and variations of uptake characteristics by individual plant species within the landuse types are not considered explicitly.

AtmosphericDeposition.z₀_table — Constant

Values for the characteristic radii of collectors [m] where the columns are land use categories and the rows are seasonal categories. Land-use categories (LUCs)

Evergreen–needleleaf trees

Deciduous broadleaf trees

Grass

Desert

Shrubs and interrupted woodlands

Seasonal categories (SC)

Midsummer with lush vegetation

Autumn with cropland not harvested

Late autumn after frost, no snow

Winter, snow on ground

Transitional

given in Seinfeld and Pandis Table 19.2

AtmosphericDeposition.DrydepositionG — Type

Description: This is a box model used to calculate the gas species concentration rate changed by dry deposition. Build Drydeposition model (gas)

Example

@parameters t 
	d = DrydepositionG(t)

AtmosphericDeposition.Wetdeposition — Type

Description: This is a box model used to calculate the concentration rate changed by wet deposition. Build Wetdeposition model

Example

@parameters t 
	wd = Wetdeposition(t)

AtmosphericDeposition.DryDepGas — Method

Function DryDepGas calculates dry deposition velocity [m/s] for a gas species, where z is the height of the surface layer [m], zo is roughness length [m], u_star is friction velocity [m/s], L is Monin-Obukhov length [m], T is surface air temperature [K], ρA is air density [kg/m3] gasData is data about the gas species for surface resistance calculations, G is solar irradiation [W m-2], Θ is the slope of the local terrain [radians], iSeason and iLandUse are indexes for the season and land use, dew and rain indicate whether there is dew or rain on the ground, and isSO2 and isO3 indicate whether the gas species of interest is either SO2 or O3, respectively. Based on Seinfeld and Pandis (2006) equation 19.2.

AtmosphericDeposition.DryDepParticle — Method

Function DryDepParticle calculates particle dry deposition velocity [m/s] where z is the height of the surface layer [m], zo is roughness length [m], u_star is friction velocity [m/s], L is Monin-Obukhov length [m], Dp is particle diameter [m], Ts is surface air temperature [K], P is pressure [Pa], ρParticle is particle density [kg/m3], ρAir is air density [kg/m3], and iSeason and iLandUse are indexes for the season and land use. Based on Seinfeld and Pandis (2006) equation 19.7.

AtmosphericDeposition.RbGas — Method

Function RbGas calculates the quasi-laminar sublayer resistance to dry deposition for a gas species [s/m], where Sc is the dimensionless Schmidt number and u_star is the friction velocity [m/s]. From Seinfeld and Pandis (2006) equation 19.17.

AtmosphericDeposition.RbParticle — Method

Function RbParticle calculates the quasi-laminar sublayer resistance to dry deposition for a particles [s/m], where Sc is the dimensionless Schmidt number, u_star is the friction velocity [m/s], St is the dimensionless Stokes number, Dp is particle diameter [m], and iSeason and iLandUse are season and land use indexes, respectively. From Seinfeld and Pandis (2006) equation 19.27.

AtmosphericDeposition.WetDeposition — Method

Calculate wet deposition based on formulas at www.emep.int/UniDoc/node12.html. Inputs are fraction of grid cell covered by clouds (cloudFrac), rain mixing ratio (qrain), air density (ρ_air [kg/m3]), and fall distance (Δz [m]). Outputs are wet deposition rates for PM2.5, SO2, and other gases (wdParticle, wdSO2, and wdOtherGas [1/s]).

AtmosphericDeposition.cc — Method

Function cc calculates the Cunnningham slip correction factor where Dp is particle diameter [m], T is temperature [K], and P is pressure [Pa]. From Seinfeld and Pandis (2006) equation 9.34.

AtmosphericDeposition.dH2O — Method

Function dH2O calculates molecular diffusivity of water vapor in air [m2/s] where T is temperature [K] using a regression fit to data in Bolz and Tuve (1976) found here: http://www.cambridge.org/us/engineering/author/nellisandklein/downloads/examples/EXAMPLE_9.2-1.pdf

AtmosphericDeposition.dParticle — Method

Function dParticle calculates the brownian diffusivity of a particle [m2/s] using the Stokes-Einstein-Sutherland relation (Seinfeld and Pandis eq. 9.73) where T is air temperature [K], P is pressure [Pa], Dp is particle diameter [m], and μ is air dynamic viscosity [kg/(s m)]

AtmosphericDeposition.mfp — Method

Function mfp calculates the mean free path of air [m] where T is temperature [K] P is pressure [Pa], and Mu is dynamic viscosity [kg/(m s)]. From Seinfeld and Pandis (2006) equation 9.6

AtmosphericDeposition.mu — Method

Function mu calculates the dynamic viscosity of air [kg m-1 s-1] where T is temperature [K].

AtmosphericDeposition.ra — Method

Function Ra calculates aerodynamic resistance to dry deposition where z is the top of the surface layer [m], z₀ is the roughness length [m], u_star is friction velocity [m/s], and L is Monin-Obukhov length [m] Based on Seinfeld and Pandis (2006) [Seinfeld, J.H. and Pandis, S.N. (2006) Atmospheric Chemistry and Physics: From Air Pollution to Climate Change. 2nd Edition, John Wiley & Sons, New York.] equation 19.13 & 19.14.

AtmosphericDeposition.sc — Method

Function sc computes the dimensionless Schmidt number, where μ is dynamic viscosity of air [kg/(s m)], ρ is air density [kg/m3], and D is the molecular diffusivity of the gas speciesof interest [m2/s]

AtmosphericDeposition.stSmooth — Method

Function stSmooth computes the dimensionless Stokes number for dry deposition of particles on smooth surfaces or surfaces with bluff roughness elements, where vs is settling velocity [m/s], u_star is friction velocity [m/s], μ is dynamic viscosity of air [kg/(s m)], and ρ is air density [kg/m3], based on Seinfeld and Pandis (2006) equation 19.23.

AtmosphericDeposition.stVeg — Method

Function stVeg computes the dimensionless Stokes number for dry deposition of particles on vegetated surfaces, where vs is settling velocity [m/s], u_star is friction velocity [m/s], and A is the characteristic collector radius [m], based on Seinfeld and Pandis (2006) equation 19.24.

AtmosphericDeposition.vs — Method

Function vs calculates the terminal setting velocity of a particle where Dp is particle diameter [m], ρₚ is particle density [kg/m3], Cc is the Cunningham slip correction factor, and μ is air dynamic viscosity [kg/(s m)]. From equation 9.42 in Seinfeld and Pandis (2006).